* Streamlined some variables to be of the same type regardless of the architecture used
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@524 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/common/knote.c
@@ -120,7 +120,7 @@ knote_delete(struct filter *filt, struct knote *kn)
 }
 
 struct knote *
-knote_lookup(struct filter *filt, short ident)
+knote_lookup(struct filter *filt, uintptr_t ident)
 {
     struct knote query;
     struct knote *ent = NULL;
